Add improvement: initialize logger on factory
