+++ NetSDK/Services/Client/Classes/JSONFileClient.cs
@@ -21,7 +21,6 @@ namespace Splitio.Services.Client.Classes
         private static readonly Logger Log = LogManager.GetLogger(typeof(JSONFileClient).ToString());
         public JSONFileClient(string splitsFilePath, string segmentsFilePath, ISegmentCache segmentCacheInstance = null, ISplitCache splitCacheInstance = null, IImpressionListener treatmentLogInstance = null, bool isLabelsEnabled = true)
         {
-            InitializeLogger();
             segmentCache = segmentCacheInstance ?? new InMemorySegmentCache(new ConcurrentDictionary<string, Segment>());
             var segmentFetcher = new JSONFileSegmentFetcher(segmentsFilePath, segmentCache);
             var splitParser = new InMemorySplitParser(segmentFetcher, segmentCache);
@@ -34,39 +33,6 @@ namespace Splitio.Services.Client.Classes
             impressionListener = treatmentLogInstance;
             splitter = new Splitter();
             LabelsEnabled = isLabelsEnabled;
-        }
-
-        private void InitializeLogger()
-        {
-            var fileTarget = new FileTarget();
-            fileTarget.Name = "splitio";
-            fileTarget.FileName = @".\Logs\splitio.log";
-            fileTarget.ArchiveFileName = @".\Logs\splitio.{#}.log";
-            fileTarget.LineEnding = LineEndingMode.CRLF;
-            fileTarget.Layout = "${longdate} ${level: uppercase = true} ${logger} - ${message} - ${exception:format=tostring}";
-            fileTarget.ConcurrentWrites = true;
-            fileTarget.CreateDirs = true;
-            fileTarget.ArchiveNumbering = ArchiveNumberingMode.DateAndSequence;
-            fileTarget.ArchiveAboveSize = 200000000;
-            fileTarget.ArchiveDateFormat = "yyyyMMdd";
-            fileTarget.MaxArchiveFiles = 30;
-            var rule = new LoggingRule("*", LogLevel.Debug, fileTarget);
-
-            if (LogManager.Configuration == null)
-            {
-                var config = new LoggingConfiguration();
-                config.AddTarget("splitio", fileTarget);
-                config.LoggingRules.Add(rule);
-                LogManager.Configuration = config;
-            }
-            else
-            {
-                if (LogManager.Configuration.ConfiguredNamedTargets.Where(x => x.Name == "splitio").FirstOrDefault() == null)
-                {
-                    LogManager.Configuration.AddTarget("splitio", fileTarget);
-                    LogManager.Configuration.LoggingRules.Add(rule);
-                }
-            }
         }
 
         public void RemoveSplitFromCache(string splitName)
