+++ NetSDK/Services/Client/Classes/LocalhostClient.cs
@@ -21,7 +21,6 @@ namespace Splitio.Services.Client.Classes
 
         public LocalhostClient(string filePath, Splitter splitter = null)
         {
-            InitializeLogger();
             fullPath = LookupFilePath(filePath);
             var directoryPath = Path.GetDirectoryName(fullPath);
 
@@ -113,39 +112,6 @@ namespace Splitio.Services.Client.Classes
             };
 
             return split;
-        }
-
-        private void InitializeLogger()
-        {
-            var fileTarget = new FileTarget();
-            fileTarget.Name = "splitio";
-            fileTarget.FileName = @".\Logs\splitio.log";
-            fileTarget.ArchiveFileName = @".\Logs\splitio.{#}.log";
-            fileTarget.LineEnding = LineEndingMode.CRLF;
-            fileTarget.Layout = "${longdate} ${level: uppercase = true} ${logger} - ${message} - ${exception:format=tostring}";
-            fileTarget.ConcurrentWrites = true;
-            fileTarget.CreateDirs = true;
-            fileTarget.ArchiveNumbering = ArchiveNumberingMode.DateAndSequence;
-            fileTarget.ArchiveAboveSize = 200000000;
-            fileTarget.ArchiveDateFormat = "yyyyMMdd";
-            fileTarget.MaxArchiveFiles = 30;
-            var rule = new LoggingRule("*", LogLevel.Debug, fileTarget);
-
-            if (LogManager.Configuration == null)
-            {
-                var config = new LoggingConfiguration();
-                config.AddTarget("splitio", fileTarget);
-                config.LoggingRules.Add(rule);
-                LogManager.Configuration = config;
-            }
-            else
-            {
-                if (LogManager.Configuration.ConfiguredNamedTargets.Where(x => x.Name == "splitio").FirstOrDefault() == null)
-                {
-                    LogManager.Configuration.AddTarget("splitio", fileTarget);
-                    LogManager.Configuration.LoggingRules.Add(rule);
-                }
-            }
         }
 
         private void BuildSplitter(Splitter splitter)
