+++ NetSDK/Services/Client/Classes/RedisClient.cs
@@ -38,7 +38,6 @@ namespace Splitio.Services.Client.Classes
 
         public RedisClient(ConfigurationOptions config)
         {
-            InitializeLogger();
             ReadConfig(config);
             BuildRedisCache();
             BuildTreatmentLog(config); 
@@ -49,39 +48,6 @@ namespace Splitio.Services.Client.Classes
         }
 
 
-        private void InitializeLogger()
-        {
-
-            var fileTarget = new FileTarget();
-            fileTarget.Name = "splitio";
-            fileTarget.FileName = @".\Logs\splitio.log";
-            fileTarget.ArchiveFileName = @".\Logs\splitio.{#}.log";
-            fileTarget.LineEnding = LineEndingMode.CRLF;
-            fileTarget.Layout = "${longdate} ${level: uppercase = true} ${logger} - ${message} - ${exception:format=tostring}";
-            fileTarget.ConcurrentWrites = true;
-            fileTarget.CreateDirs = true;
-            fileTarget.ArchiveNumbering = ArchiveNumberingMode.DateAndSequence;
-            fileTarget.ArchiveAboveSize = 200000000;
-            fileTarget.ArchiveDateFormat = "yyyyMMdd";
-            fileTarget.MaxArchiveFiles = 30;
-            var rule = new LoggingRule("*", LogLevel.Debug, fileTarget);
-
-            if (LogManager.Configuration == null)
-            {
-                var config = new LoggingConfiguration();
-                config.AddTarget("splitio", fileTarget);
-                config.LoggingRules.Add(rule);
-                LogManager.Configuration = config;
-            }
-            else
-            {
-                if (LogManager.Configuration.ConfiguredNamedTargets.Where(x => x.Name == "splitio").FirstOrDefault() == null)
-                {
-                    LogManager.Configuration.AddTarget("splitio", fileTarget);
-                    LogManager.Configuration.LoggingRules.Add(rule);
-                }
-            }
-        }
         private void ReadConfig(ConfigurationOptions config)
         {
             SdkVersion = ".NET-" + Version.SplitSdkVersion;
