+++ NetSDK/Services/Client/Classes/SelfRefreshingClient.cs
@@ -54,7 +54,6 @@ namespace Splitio.Services.Client.Classes
         /// </summary>
         private const int InitialCapacity = 31;
 
-
         private IReadinessGatesCache gates;
         private SelfRefreshingSplitFetcher splitFetcher;
         private ISplitSdkApiClient splitSdkApiClient;
@@ -66,7 +65,6 @@ namespace Splitio.Services.Client.Classes
 
         public SelfRefreshingClient(string apiKey, ConfigurationOptions config)
         {
-            InitializeLogger();
             ApiKey = apiKey;
             ReadConfig(config);
             BuildSdkReadinessGates();
@@ -160,39 +158,6 @@ namespace Splitio.Services.Client.Classes
             ((SelfRefreshingSplitFetcher)splitFetcher).Stop();
             ((SelfRefreshingSegmentFetcher)selfRefreshingSegmentFetcher).Stop();
             ((SelfUpdatingTreatmentLog)treatmentLog).Stop();
-        }
-
-        private void InitializeLogger()
-        {
-            var fileTarget = new FileTarget();
-            fileTarget.Name = "splitio";
-            fileTarget.FileName = @".\Logs\splitio.log";
-            fileTarget.ArchiveFileName = @".\Logs\splitio.{#}.log";
-            fileTarget.LineEnding = LineEndingMode.CRLF;
-            fileTarget.Layout = "${longdate} ${level: uppercase = true} ${logger} - ${message} - ${exception:format=tostring}";
-            fileTarget.ConcurrentWrites = true;
-            fileTarget.CreateDirs = true;
-            fileTarget.ArchiveNumbering = ArchiveNumberingMode.DateAndSequence;
-            fileTarget.ArchiveAboveSize = 200000000;
-            fileTarget.ArchiveDateFormat = "yyyyMMdd";
-            fileTarget.MaxArchiveFiles = 30;
-            var rule = new LoggingRule("*", LogLevel.Debug, fileTarget);
-
-            if (LogManager.Configuration == null)
-            {
-                var config = new LoggingConfiguration();
-                config.AddTarget("splitio", fileTarget);
-                config.LoggingRules.Add(rule);
-                LogManager.Configuration = config;
-            }
-            else
-            {
-                if (LogManager.Configuration.ConfiguredNamedTargets.Where(x => x.Name == "splitio").FirstOrDefault() == null)
-                {
-                    LogManager.Configuration.AddTarget("splitio", fileTarget);
-                    LogManager.Configuration.LoggingRules.Add(rule);
-                }
-            }
         }
 
         private void BuildSplitter()
