+++ NetSDK/Services/Client/Classes/SplitFactory.cs
@@ -1,5 +1,9 @@
-﻿using Splitio.Services.Client.Interfaces;
+﻿using NLog;
+using NLog.Config;
+using NLog.Targets;
+using Splitio.Services.Client.Interfaces;
 using System;
+using System.Linq;
 
 namespace Splitio.Services.Client.Classes
 {
@@ -14,6 +18,40 @@ namespace Splitio.Services.Client.Classes
         {
             this.apiKey = apiKey;
             this.options = options;
+            InitializeLogger();
+        }
+
+        private void InitializeLogger()
+        {
+            var fileTarget = new FileTarget();
+            fileTarget.Name = "splitio";
+            fileTarget.FileName = @".\Logs\splitio.log";
+            fileTarget.ArchiveFileName = @".\Logs\splitio.{#}.log";
+            fileTarget.LineEnding = LineEndingMode.CRLF;
+            fileTarget.Layout = "${longdate} ${level: uppercase = true} ${logger} - ${message} - ${exception:format=tostring}";
+            fileTarget.ConcurrentWrites = true;
+            fileTarget.CreateDirs = true;
+            fileTarget.ArchiveNumbering = ArchiveNumberingMode.DateAndSequence;
+            fileTarget.ArchiveAboveSize = 200000000;
+            fileTarget.ArchiveDateFormat = "yyyyMMdd";
+            fileTarget.MaxArchiveFiles = 30;
+            var rule = new LoggingRule("*", LogLevel.Debug, fileTarget);
+
+            if (LogManager.Configuration == null)
+            {
+                var config = new LoggingConfiguration();
+                config.AddTarget("splitio", fileTarget);
+                config.LoggingRules.Add(rule);
+                LogManager.Configuration = config;
+            }
+            else
+            {
+                if (LogManager.Configuration.ConfiguredNamedTargets.Where(x => x.Name == "splitio").FirstOrDefault() == null)
+                {
+                    LogManager.Configuration.AddTarget("splitio", fileTarget);
+                    LogManager.Configuration.LoggingRules.Add(rule);
+                }
+            }
         }
 
         public ISplitClient Client()
