+++ NetSDK/Version.cs
@@ -3,7 +3,7 @@ namespace Splitio
 {
     public static class Version
     {
-        public static string SplitSdkVersion = "3.1.2-rc179";
+        public static string SplitSdkVersion = "3.1.2-rc188";
         public static string SplitSpecVersion = "1.0";
     }
 }
